+++ CHANGES.rst
@@ -49,6 +49,8 @@ Bug fixes
 * If an arbitrary GAP error occurs while looking up a global variable with
   ``Gap.__getattr__`` it is handled and re-raised as an ``AttributeError``.
 
+* The ``Gap.__repr__`` method displays names of subclasses correctly.
+
 
 v0.1.0a2 (2021-02-03)
 ---------------------